Breve refactoring codice, aggiunta documentazione e commenti, aggiunto catch per KeyboardInterrupt con attesa per completare i download in corso, rimosse variabili inutili
