+++ README.md
@@ -1,2 +1,5 @@
 # doExercises
 Script python per il download degli esercizi della piattaforma DoExercises (corso di probabilità e statistica)
+
+### Utilizzo
+Usare `python doexercises.py --help` ottenere informazioni sui comandi possibili
